refactor(mcp/tools): extract request/argument types to requests.rs
tools/mod.rs fused two duties: the per-tool request/argument types (the *Args
structs, *Include and Mcp* mode enums with their tolerant Deserialize impls and
as_str/parse/core conversions) and the call dispatch. Move the types into a
sibling requests.rs, leaving mod.rs as dispatch (call_tool / call_tool_for_config
/ finalize_tool_result), the read-only classifier, and the include-flag helper.

mod.rs drops 833 -> 118 lines. The conversion methods crossing into handlers
(McpGraphMode::as_str, McpGraphResolutionMode::core, McpGraphEdgeKind::as_str,
McpRepoBriefMode::core, Memory{Create,Update}Args::core) widen to pub(super);
the rest stay private. mod.rs keeps the shared import set the sibling modules
reach through `use super::*`. No behavior change.
